refactor(stellar): extract crypto ops to dedicated module
