+++ contracts/stellar/protocol/src/instructions/delegation.rs
@@ -1,5 +1,6 @@
-use soroban_sdk::{Address, Env, BytesN, Bytes, Vec};
-use crate::state::{DataKey, Attestation, DelegatedAttestationRequest, DelegatedRevocationRequest, BlsPublicKey};
+use soroban_sdk::{Address, Env, BytesN, Bytes};
+use crate::instructions::verify_bls_signature;
+use crate::state::{DataKey, Attestation, DelegatedAttestationRequest, DelegatedRevocationRequest};
 use crate::errors::Error;
 use crate::utils;
 use crate::events;
@@ -371,271 +372,4 @@ fn create_revocation_message(
     // Return hash of the complete message
     env.crypto().sha256(&message).into()
 }
-
-/// **CRITICAL CRYPTOGRAPHIC FUNCTION**: Verifies BLS12-381 signature for delegated attestation
-///
-/// This is the core cryptographic security function that validates signatures created off-chain.
-/// It implements the BLS signature verification algorithm to prove that the attester's private
-/// key was used to sign the specific message hash. This is the primary defense against
-/// unauthorized attestation creation and signature forgery.
-///
-/// # Cryptographic Model
-/// BLS (Boneh-Lynn-Shacham) signatures provide:
-/// - **Short Signatures**: 96-byte signatures vs 64-byte ECDSA
-/// - **Aggregation Support**: Multiple signatures can be combined
-/// - **Quantum Resistance**: Based on discrete log problem in elliptic curve groups
-/// - **Deterministic**: Same message + private key always produces same signature
-///
-/// # Security Properties
-/// - **Unforgeable**: Without private key, cannot create valid signatures
-/// - **Non-Malleable**: Cannot modify signature to create new valid signature  
-/// - **Message Binding**: Signature is cryptographically bound to exact message
-/// - **Key Authenticity**: Signature proves possession of private key without revealing it
-///
-/// # BLS Signature Verification Process
-/// 1. **Key Lookup**: Retrieve attester's registered G2 public key
-/// 2. **Message Hashing**: Hash message to G1 point on elliptic curve
-/// 3. **Signature Parsing**: Convert 96-byte signature to G1 point
-/// 4. **Pairing Check**: Verify e(H(m), PK) = e(σ, G2) where:
-///    - H(m) = hashed message (G1 point)
-///    - PK = public key (G2 point)  
-///    - σ = signature (G1 point)
-///    - G2 = generator of G2 group
-///
-/// # Parameters
-/// * `env` - Soroban environment for cryptographic operations
-/// * `message` - SHA256 hash of the signed message (32 bytes)
-/// * `signature` - BLS12-381 signature in compressed G1 format (96 bytes)
-/// * `attester` - Wallet address of the signer (for public key lookup)
-///
-/// # Returns
-/// * `Ok(())` - Signature is cryptographically valid
-/// * `Err(Error::InvalidSignature)` - Signature verification failed (forge attempt or no key)
-///
-/// # Critical Security Checks
-/// 1. **Key Registration**: Attester must have registered BLS public key
-/// 2. **Cryptographic Verification**: Pairing equation must hold
-/// 3. **Point Validation**: All elliptic curve points must be valid
-/// 4. **Domain Separation**: Uses standard BLS signature DSTs
-///
-/// # Attack Vectors & Mitigations
-/// * **Signature Forgery**: Creating signatures without private key
-///   - *Mitigation*: Cryptographically impossible due to discrete log problem
-/// * **Key Substitution**: Using different public key than registered
-///   - *Mitigation*: Public key lookup tied to specific attester address
-/// * **Message Manipulation**: Changing message after signing
-///   - *Mitigation*: Signature cryptographically bound to exact message hash
-/// * **Replay Attacks**: Reusing valid signatures
-///   - *Mitigation*: Nonce system prevents signature reuse (handled separately)
-/// * **Invalid Point Attacks**: Malformed elliptic curve points
-///   - *Mitigation*: Point validation in cryptographic library
-///
-/// # Implementation Status
-/// **IMPORTANT**: This is a proof-of-concept implementation for testing BLS API.
-/// Production deployment requires:
-/// 1. **Proper Point Deserialization**: Convert compressed bytes to curve points
-/// 2. **Standard DST Values**: Use official BLS signature domain separation tags
-/// 3. **G2 Generator**: Use correct G2 generator point from BLS12-381 spec
-/// 4. **Error Handling**: Comprehensive validation of all cryptographic operations
-///
-/// # Cross-Platform Compatibility
-/// Must be compatible with @noble/curves BLS implementation:
-/// ```javascript
-/// import { bls12_381 } from '@noble/curves/bls12-381';
-/// 
-/// // Off-chain signature creation
-/// const signature = bls12_381.sign(messageHash, privateKey);
-/// 
-/// // This Rust function must verify signatures created by above
-/// ```
-///
-/// # Q/A Testing Focus
-/// 1. **Signature Verification**: Test with known test vectors from BLS12-381 spec
-/// 2. **Invalid Signature Rejection**: Verify random signatures are rejected
-/// 3. **Key Mismatch**: Test signatures with wrong public keys fail
-/// 4. **Message Tampering**: Verify changed messages fail verification  
-/// 5. **Cross-Platform**: Test JavaScript-created signatures verify in Rust
-/// 6. **Edge Cases**: Test with invalid/malformed signature bytes
-/// 7. **Performance**: Measure verification time for DoS analysis
-fn verify_bls_signature(
-    env: &Env,
-    message: &BytesN<32>,
-    signature: &BytesN<96>,
-    attester: &Address,
-) -> Result<(), Error> {
-    // STEP 1: Look up the attester's registered BLS public key
-    // Each attester must register exactly one immutable BLS key
-    let pk_key = DataKey::AttesterPublicKey(attester.clone());
-    let bls_key = env.storage().persistent()
-        .get::<DataKey, BlsPublicKey>(&pk_key)
-        .ok_or(Error::InvalidSignature)?; // No key registered = cannot verify
-    
-    // STEP 2: Initialize BLS12-381 cryptographic operations
-    let bls = env.crypto().bls12_381();
-    
-    // STEP 3: Hash message to G1 point using standard BLS DST
-    // This creates the point H(m) that was signed by the private key
-    let message_bytes = Bytes::from_array(env, &message.to_array());
-    let dst = Bytes::from_slice(env, b"BLS_SIG_BLS12381G1_XMD:SHA-256_SSWU_RO_NUL_");
-    let h_m = bls.hash_to_g1(&message_bytes, &dst);
-    
-    // STEP 4: Parse signature from compressed bytes to G1 point
-    // NOTE: This is simplified for proof-of-concept
-    // Production requires proper point deserialization validation
-    let sig_as_bytes = Bytes::from_array(env, &signature.to_array());
-    let sig_point = bls.hash_to_g1(&sig_as_bytes, &dst); // TODO: Use proper parsing
-    
-    // STEP 5: Parse public key from compressed bytes to G2 point
-    // The registered public key is used to verify signature authenticity
-    let pk_as_bytes = Bytes::from_array(env, &bls_key.key.to_array());
-    let dst_g2 = Bytes::from_slice(env, b"BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_NUL_");
-    let pk_point = bls.hash_to_g2(&pk_as_bytes, &dst_g2); // TODO: Use proper parsing
-    
-    // STEP 6: Get G2 generator point for pairing verification
-    // NOTE: This is placeholder - production needs actual BLS12-381 G2 generator
-    let g2_gen = bls.hash_to_g2(&Bytes::from_slice(env, b"BLS12381G2_GENERATOR"), &dst_g2);
-    
-    // STEP 7: Perform pairing check to verify signature
-    // Verifies: e(H(m), PK) = e(σ, G2)
-    // This cryptographically proves signature was created with corresponding private key
-    let g1_points = Vec::from_array(env, [h_m, sig_point]);
-    let g2_points = Vec::from_array(env, [pk_point, g2_gen]);
-    
-    let pairing_result = bls.pairing_check(g1_points, g2_points);
-    
-    if pairing_result {
-        Ok(()) // Signature is cryptographically valid
-    } else {
-        Err(Error::InvalidSignature) // Signature verification failed
-    }
-}
-
-/// Gets the next nonce for an attester.
-///
-/// # Arguments
-/// * `env` - The Soroban environment
-/// * `attester` - The address of the attester
-///
-/// # Returns
-/// * `u64` - The next nonce to be used
-pub fn get_next_nonce(env: &Env, attester: &Address) -> u64 {
-    let nonce_key = DataKey::AttesterNonce(attester.clone());
-    env.storage().persistent()
-        .get::<DataKey, u64>(&nonce_key)
-        .unwrap_or(0)
-}
-
-/// Registers a BLS public key for an attester.
-///
-/// Each wallet address can register exactly one BLS public key.
-/// Once registered, the key is immutable - cannot be updated or revoked.
-/// To use a different key, use a different wallet address.
-///
-/// # Arguments
-/// * `env` - The Soroban environment
-/// * `attester` - The address of the attester registering the key
-/// * `public_key` - The BLS12-381 G2 public key (96 bytes)
-///
-/// # Returns
-/// * `Result<(), Error>` - Success or error (fails if key already exists)
-pub fn register_bls_public_key(
-    env: &Env,
-    attester: Address,
-    public_key: BytesN<96>,
-) -> Result<(), Error> {
-    attester.require_auth();
-    
-    let pk_key = DataKey::AttesterPublicKey(attester.clone());
-    
-    // Check if this address already has a key registered
-    if env.storage().persistent().has(&pk_key) {
-        // Key already registered - immutable, cannot update
-        return Err(Error::AlreadyInitialized);
-    }
-    
-    let timestamp = env.ledger().timestamp();
-    let bls_key = BlsPublicKey {
-        key: public_key.clone(),
-        registered_at: timestamp,
-    };
-    
-    // Store the key (one per address, immutable)
-    env.storage().persistent().set(&pk_key, &bls_key);
-    
-    // Emit registration event
-    crate::events::publish_bls_key_registered(
-        env,
-        &attester,
-        &public_key,
-        timestamp,
-    );
-    
-    Ok(())
-}
-
-/// Gets the BLS public key for an attester.
-///
-/// # Arguments
-/// * `env` - The Soroban environment
-/// * `attester` - The address of the attester
-///
-/// # Returns
-/// * `Option<BlsPublicKey>` - The public key if registered
-pub fn get_bls_public_key(env: &Env, attester: &Address) -> Option<BlsPublicKey> {
-    let pk_key = DataKey::AttesterPublicKey(attester.clone());
-    env.storage().persistent().get(&pk_key)
-}
-
-/*
-JavaScript Integration Guide for BLS12-381 Signatures
-====================================================
-
-IMPORTANT: The attester (entity making claims) creates signatures, NOT the subject.
-The subject being attested never needs to interact with the blockchain.
-
-To create compatible signatures using @noble/curves:
-
-```javascript
-import { bls12_381 } from '@noble/curves/bls12-381';
-import { sha256 } from '@noble/hashes/sha256';
-
-// 1. Attester generates their keypair (done once)
-const attesterPrivateKey = bls12_381.utils.randomPrivateKey();
-const attesterPublicKey = bls12_381.getPublicKey(attesterPrivateKey);
-
-// 2. Create attestation message (must match Rust implementation)
-function createAttestationMessage(request) {
-    const domainSeparator = new TextEncoder().encode("ATTEST_PROTOCOL_V1_DELEGATED");
-    const schemaBytes = new Uint8Array(request.schema_uid);
-    const nonceBytes = new DataView(new ArrayBuffer(8));
-    nonceBytes.setBigUint64(0, BigInt(request.nonce), false); // big-endian
-    const deadlineBytes = new DataView(new ArrayBuffer(8));
-    deadlineBytes.setBigUint64(0, BigInt(request.deadline), false);
-    
-    // Concatenate all fields
-    const message = new Uint8Array([
-        ...domainSeparator,
-        ...schemaBytes,
-        ...new Uint8Array(nonceBytes.buffer),
-        ...new Uint8Array(deadlineBytes.buffer),
-        // Add other fields as needed
-    ]);
-    
-    return sha256(message);
-}
-
-// 3. Attester signs the message about a subject
-const message = createAttestationMessage(attestationRequest);
-const signature = bls12_381.sign(message, attesterPrivateKey);
-
-// 4. The signature can be submitted by ANYONE
-// The attester doesn't need to pay gas fees
-// The subject never needs to sign or interact
-```
-
-Domain Separation Tags (DST):
-- Attestation messages: "ATTEST_PROTOCOL_V1_DELEGATED"  
-- Revocation messages: "REVOKE_PROTOCOL_V1_DELEGATED"
-- BLS G1 hashing: "BLS_SIG_BLS12381G1_XMD:SHA-256_SSWU_RO_NUL_"
-- BLS G2 hashing: "BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_NUL_"
-*/
+
