+++ contracts/stellar/protocol/src/instructions/mod.rs
@@ -1,13 +1,17 @@
 pub mod schema;
 pub mod attest;
 pub mod delegation;
+pub mod crypto;
 
 pub use schema::{register_schema};
 pub use attest::{attest, get_attestation, revoke_attestation, list_attestations};
 pub use delegation::{
     attest_by_delegation, 
-    revoke_by_delegation, 
-    get_next_nonce, 
-    register_bls_public_key, 
+    revoke_by_delegation
+};
+
+pub use crypto::{
+    verify_bls_signature,
+    register_bls_public_key,
     get_bls_public_key
 };
